rename Footer
