+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import Navbar from "@/components/Navbar";
-import Footer from "@/components/Footer-test";
+import Footer from "@/components/Footer";
 import "./globals.css";
 
 const geistSans = Geist({
